requirements.txt
Pillow>=9.0.0
sv-ttk>=2.6.0
pyperclip>=1.8.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.0